added pagination in rooms api

+++ backend/src/main/java/at/teamhasi/hotel/dto/PaginationDto.java
@@ -1,6 +1,5 @@
 package at.teamhasi.hotel.dto;
 
-import at.teamhasi.hotel.entities.MediaEntity;
 import lombok.*;
 
 @Getter
@@ -12,5 +11,5 @@ public class PaginationDto {
     private int limit;
     private int offset;
     private Integer nextOffset;
-    private int totalCount;
+    private Long totalCount;
 }
